# services/api/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactView": {
        "properties": {
          "byte_length": {
            "title": "Byte Length",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "sha256",
          "byte_length",
          "created_at"
        ],
        "title": "ArtifactView",
        "type": "object"
      },
...
      "ExecutionView": {
        "properties": {
          "attempt_number": {
            "title": "Attempt Number",
            "type": "integer"
          },
          "backend": {
            "enum": [
              "REPLAY",
              "SOLARI"
            ],
            "title": "Backend",
            "type": "string"
          },
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          },
          "cleanup_state": {
            "enum": [
              "PASS",
              "FAIL",
              "NOT_REQUIRED"
            ],
            "title": "Cleanup State",
            "type": "string"
          },
          "command_sha256": {
            "title": "Command Sha256",
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "evidence": {
            "$ref": "#/components/schemas/ArtifactView"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finding_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "infrastructure_state": {
            "enum": [
              "PASS",
              "FAIL"
            ],
            "title": "Infrastructure State",
            "type": "string"
          },
          "infrastructure_step": {
            "title": "Infrastructure Step",
            "type": "string"
          },
          "language": {
            "enum": [
              "python",
              "typescript"
            ],
            "title": "Language",
            "type": "string"
          },
          "output_truncated": {
            "title": "Output Truncated",
            "type": "boolean"
          },
          "package_name": {
            "title": "Package Name",
            "type": "string"
          },
          "package_version": {
            "title": "Package Version",
            "type": "string"
          },
          "phase": {
            "enum": [
              "VERIFY",
              "FIX_VERIFY"
            ],
            "title": "Phase",
            "type": "string"
          },
          "proposal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Id"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          },
          "source_surface": {
            "title": "Source Surface",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "stderr": {
            "title": "Stderr",
            "type": "string"
          },
          "stdout": {
            "title": "Stdout",
            "type": "string"
          },
          "subject_state": {
            "enum": [
              "PASS",
              "FAIL",
              "NOT_RUN"
            ],
            "title": "Subject State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "finding_id",
          "proposal_id",
          "attempt_number",
          "language",
          "source_surface",
          "phase",
          "backend",
          "infrastructure_state",
          "infrastructure_step",
          "subject_state",
          "sandbox_id",
          "package_name",
          "package_version",
          "source_path",
          "source_sha256",
          "command_sha256",
          "exit_code",
          "stdout",
          "stderr",
          "output_truncated",
          "cleanup_state",
          "cancelled",
          "error_code",
          "duration_ms",
          "started_at",
          "completed_at",
          "evidence"
        ],
        "title": "ExecutionView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposalList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProposalView"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProposalList",
        "type": "object"
      },
      "ProposalView": {
        "properties": {
          "changed_lines": {
            "title": "Changed Lines",
            "type": "integer"
          },
          "checkout_modified": {
            "const": false,
            "default": false,
            "title": "Checkout Modified",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finding_id": {
            "format": "uuid",
            "title": "Finding Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patch": {
            "title": "Patch",
            "type": "string"
          },
          "patch_artifact": {
            "$ref": "#/components/schemas/ArtifactView"
          },
          "proposed_artifact": {
            "$ref": "#/components/schemas/ArtifactView"
          },
          "proposed_sha256": {
            "title": "Proposed Sha256",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_artifact": {
            "$ref": "#/components/schemas/ArtifactView"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "source_sha256": {
            "title": "Source Sha256",
            "type": "string"
          },
          "source_surface": {
            "enum": [
              "python",
              "docs_python"
            ],
            "title": "Source Surface",
            "type": "string"
          },
          "state": {
            "enum": [
              "FIX_PROPOSED",
              "FIX_VERIFIED",
              "DISMISSED"
            ],
            "title": "State",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification_job_state": {
            "anyOf": [
              {
                "enum": [
                  "QUEUED",
                  "LEASED",
                  "COMPLETED",
                  "FAILED",
                  "CANCELLED"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Job State"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "run_id",
          "finding_id",
          "state",
          "source_surface",
          "source_path",
          "source_sha256",
          "proposed_sha256",
          "changed_lines",
          "patch",
          "created_at",
          "verified_at",
          "verification_job_state",
          "verification",
          "source_artifact",
          "proposed_artifact",
          "patch_artifact"
        ],
        "title": "ProposalView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Noxyn finds API ecosystem drift; Solari proves it.",
    "title": "Noxyn-Solari API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/findings/{finding_id}/proposals": {
      "post": {
        "operationId": "createFindingProposal",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Noxyn-Test-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Noxyn-Test-User"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Proposal",
        "tags": [
          "fix proposals"
        ]
      }
    },
...
    "/v1/proposals/{proposal_id}": {
      "get": {
        "operationId": "getProposal",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Noxyn-Test-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Noxyn-Test-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Proposal",
        "tags": [
          "fix proposals"
        ]
      }
    },
    "/v1/proposals/{proposal_id}/verify": {
      "post": {
        "operationId": "verifyProposal",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Noxyn-Test-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Noxyn-Test-User"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Proposal",
        "tags": [
          "fix proposals"
        ]
      }
    },
...
    "/v1/runs/{run_id}/proposals": {
      "get": {
        "operationId": "listRunProposals",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Noxyn-Test-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Noxyn-Test-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals",
        "tags": [
          "fix proposals"
        ]
      }
    }
  }
}